requirements.txt
IPy==0.83
redis==2.10.6
python-libnmap==0.7.0